@@ -18,7 +18,7 @@
  * @license http://opensource.org/licenses/MIT The MIT License
  * @version 2.1
  */
-class PicoTwigExtension extends Twig_Extension
+class PicoTwigExtension extends \Twig\Extension\AbstractExtension
 {
     /**
      * Current instance of Pico
@@ -53,7 +53,7 @@ class PicoTwigExtension extends Twig_Extension
     /**
      * Returns the name of the extension
      *
-     * @see Twig_ExtensionInterface::getName()
+     * @see \Twig\Extension\AbstractExtension::getName()
      *
      * @return string the extension name
      */
@@ -65,38 +65,38 @@ class PicoTwigExtension extends Twig_Extension
     /**
      * Returns a list of Pico-specific Twig filters
      *
-     * @see Twig_ExtensionInterface::getFilters()
+     * @see \Twig\Extension\ExtensionInterface::getFilters()
      *
-     * @return Twig_SimpleFilter[] array of Pico's Twig filters
+     * @return \Twig\TwigFilter[] array of Pico's Twig filters
      */
     public function getFilters()
     {
         return array(
-            'markdown' => new Twig_SimpleFilter(
+            'markdown' => new \Twig\TwigFilter(
                 'markdown',
                 array($this, 'markdownFilter'),
                 array('is_safe' => array('html'))
             ),
-            'map' => new Twig_SimpleFilter('map', array($this, 'mapFilter')),
-            'sort_by' => new Twig_SimpleFilter('sort_by', array($this, 'sortByFilter')),
-            'link' => new Twig_SimpleFilter('link', array($this->pico, 'getPageUrl')),
-            'url' => new Twig_SimpleFilter('url', array($this->pico, 'substituteUrl'))
+            'map' => new \Twig\TwigFilter('map', array($this, 'mapFilter')),
+            'sort_by' => new \Twig\TwigFilter('sort_by', array($this, 'sortByFilter')),
+            'link' => new \Twig\TwigFilter('link', array($this->pico, 'getPageUrl')),
+            'url' => new \Twig\TwigFilter('url', array($this->pico, 'substituteUrl'))
         );
     }
 
     /**
      * Returns a list of Pico-specific Twig functions
      *
-     * @see Twig_ExtensionInterface::getFunctions()
+     * @see \Twig\Extension\ExtensionInterface::getFunctions()
      *
-     * @return Twig_SimpleFunction[] array of Pico's Twig functions
+     * @return \Twig\TwigFunction[] array of Pico's Twig functions
      */
     public function getFunctions()
     {
         return array(
-            'url_param' => new Twig_SimpleFunction('url_param', array($this, 'urlParamFunction')),
-            'form_param' => new Twig_SimpleFunction('form_param', array($this, 'formParamFunction')),
-            'pages' => new Twig_SimpleFunction('pages', array($this, 'pagesFunction'))
+            'url_param' => new \Twig\TwigFunction('url_param', array($this, 'urlParamFunction')),
+            'form_param' => new \Twig\TwigFunction('form_param', array($this, 'formParamFunction')),
+            'pages' => new \Twig\TwigFunction('pages', array($this, 'pagesFunction'))
         );
     }
 
@@ -136,12 +136,12 @@ class PicoTwigExtension extends Twig_Extension
      *
      * @return array mapped values
      *
-     * @throws Twig_Error_Runtime
+     * @throws \Twig\Error\RuntimeError
      */
     public function mapFilter($var, $mapKeyPath)
     {
         if (!is_array($var) && (!is_object($var) || !($var instanceof Traversable))) {
-            throw new Twig_Error_Runtime(sprintf(
+            throw new \Twig\Error\RuntimeError(sprintf(
                 'The map filter only works with arrays or "Traversable", got "%s"',
                 is_object($var) ? get_class($var) : gettype($var)
             ));
@@ -178,20 +178,20 @@ class PicoTwigExtension extends Twig_Extension
      *
      * @return array sorted array
      *
-     * @throws Twig_Error_Runtime
+     * @throws \Twig\Error\RuntimeError
      */
     public function sortByFilter($var, $sortKeyPath, $fallback = 'bottom')
     {
         if (is_object($var) && ($var instanceof Traversable)) {
             $var = iterator_to_array($var, true);
         } elseif (!is_array($var)) {
-            throw new Twig_Error_Runtime(sprintf(
+            throw new \Twig\Error\RuntimeError(sprintf(
                 'The sort_by filter only works with arrays or "Traversable", got "%s"',
                 is_object($var) ? get_class($var) : gettype($var)
             ));
         }
         if (($fallback !== 'top') && ($fallback !== 'bottom') && ($fallback !== 'keep') && ($fallback !== "remove")) {
-            throw new Twig_Error_Runtime(
+            throw new \Twig\Error\RuntimeError(
                 'The sort_by filter only supports the "top", "bottom", "keep" and "remove" fallbacks'
             );
         }
@@ -399,7 +399,7 @@ class PicoTwigExtension extends Twig_Extension
      * returns Pico's full pages array.
      *
      * If `$depth` is negative after taking `$offset` into consideration, the
-     * function will throw a {@see Twig_Error_Runtime} exception, since this
+     * function will throw a {@see \Twig\Error\RuntimeError} exception, since this
      * would simply make no sense and is likely an error. Passing a negative
      * `$depthOffset` is equivalent to passing `$depthOffset = 0`.
      *
@@ -421,7 +421,7 @@ class PicoTwigExtension extends Twig_Extension
      *
      * @return array[] the data of the matched pages
      *
-     * @throws Twig_Error_Runtime
+     * @throws \Twig\Error\RuntimeError
      */
     public function pagesFunction($start = '', $depth = 0, $depthOffset = 0, $offset = 1)
     {
@@ -443,7 +443,7 @@ class PicoTwigExtension extends Twig_Extension
         $depthOffset = $depthOffset + $offset;
 
         if (($depth !== null) && ($depth < 0)) {
-            throw new Twig_Error_Runtime('The pages function doesn\'t support negative depths');
+            throw new \Twig\Error\RuntimeError('The pages function doesn\'t support negative depths');
         }
 
         $pageTree = $this->getPico()->getPageTree();
